+++ src/index.ts
@@ -63,7 +63,6 @@ export {
   SmartContract,
   method,
   declareMethods,
-  Account,
   Reducer,
 } from './lib/mina/zkapp.js';
 export { state, State, declareState } from './lib/mina/state.js';
@@ -71,6 +70,7 @@ export { state, State, declareState } from './lib/mina/state.js';
 export type { JsonProof } from './lib/proof-system/zkprogram.js';
 export {
   Proof,
+  DynamicProof,
   SelfProof,
   verify,
   Empty,
@@ -80,6 +80,7 @@ export {
 } from './lib/proof-system/zkprogram.js';
 export { Cache, CacheHeader } from './lib/proof-system/cache.js';
 
+export { Account } from './lib/mina/account.js';
 export {
   TokenId,
   AccountUpdate,